P6: Adding Color (pgs 788-790).
Seeing "The referenced script on this Behaviour (Game Object 'Scoreboard')
is missing!" error messages at runtime. They have probably been there
since scoring was added. Possibly due to the change in the name of a
script attached to a prefab.

Also had to switch to the supplied shader before colors would show up.